+++ tests/plot_test.py
@@ -172,7 +172,8 @@ class TestPlot(object):
         samples = [[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]]*21
         samples = [np.random.random(21).tolist() for i in range(21)]
         params = list(latex_labels.keys())
-        fig = gwplot._make_source_corner_plot(samples, params, latex_labels) 
+        samples = {i: j for i, j in zip(params, samples)}
+        fig = gwplot._make_source_corner_plot(samples, latex_labels) 
         assert isinstance(fig, matplotlib.figure.Figure) == True
     
     def test_extrinsic_corner_plot(self):
@@ -190,7 +191,8 @@ class TestPlot(object):
         samples = [[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21]]*21
         samples = [np.random.random(21).tolist() for i in range(21)]
         params = list(latex_labels.keys())
-        fig = gwplot._make_extrinsic_corner_plot(samples, params, latex_labels) 
+        samples = {i: j for i, j in zip(params, samples)}
+        fig = gwplot._make_extrinsic_corner_plot(samples, latex_labels) 
         assert isinstance(fig, matplotlib.figure.Figure) == True
 
     def test_sensitivity_plot(self):