+++ tests/read_test.py
@@ -123,8 +123,8 @@ class GWBaseRead(BaseRead):
         super(GWBaseRead, self).test_injection_parameters(true)
         self.result.add_injection_parameters_from_file("./tests/main_injection.xml")
         true = {
-            'mass_1': 53.333333, 'mass_2': 26.666667, 'a_1': float('nan'),
-            'a_2': float('nan'), 'tilt_1': float('nan'), 'tilt_2': float('nan'),
+            'mass_1': 53.333333, 'mass_2': 26.666667, 'a_1': 0,
+            'a_2': 0, 'tilt_1': float('nan'), 'tilt_2': float('nan'),
             'phi_jl': float('nan'), 'phi_12': float('nan'), 'psi': 1.75,
             'theta_jn': float('nan'), 'ra': float('nan'), 'dec': 1.949725,
             'luminosity_distance': 139.76429, 'geocent_time': float('nan'),